+++ src/MinFraud/Validation/Rules/Transaction.php
@@ -14,18 +14,23 @@ class Transaction extends AbstractWrapper
 {
     public function __construct()
     {
-        parent::__construct(v::keySet(
-            v::key('account', new Account(), false),
-            v::key('billing', new Billing(), false),
-            v::key('credit_card', new CreditCard(), false),
-            v::key('custom_inputs', new CustomInputs(), false),
-            v::key('device', new Device(), true),
-            v::key('email', new Email(), false),
-            v::key('event', new Event(), false),
-            v::key('order', new Order(), false),
-            v::key('payment', new Payment(), false),
-            v::key('shipping', new Shipping(), false),
-            v::key('shopping_cart', v::arrayVal()->each(new ShoppingCartItem()))
-        ));
+        parent::__construct(
+            v::allOf(
+                v::keySet(
+                    v::key('account', new Account(), false),
+                    v::key('billing', new Billing(), false),
+                    v::key('credit_card', new CreditCard(), false),
+                    v::key('custom_inputs', new CustomInputs(), false),
+                    v::key('device', new Device(), false),
+                    v::key('email', new Email(), false),
+                    v::key('event', new Event(), false),
+                    v::key('order', new Order(), false),
+                    v::key('payment', new Payment(), false),
+                    v::key('shipping', new Shipping(), false),
+                    v::key('shopping_cart', v::arrayVal()->each(new ShoppingCartItem()), false)
+                ),
+                v::arrayVal()->length(1, null),
+            ),
+        );
     }
 }